@@ -1,44 +1,49 @@
 {
-  "name": "ElectronNetAngular",
+  "name": "dotnet-angular-electron",
   "private": true,
   "version": "0.0.0",
   "scripts": {
-    "test": "karma start ClientApp/test/karma.conf.js"
+    "start": "node node_modules/webpack/bin/webpack.js --config webpack.config.vendor.js && node node_modules/webpack/bin/webpack.js --watch",
+    "lint": "node node_modules/tslint/bin/tslint -p tsconfig.json -c tslint.json",
+    "test": "karma start ClientApp/test/karma.conf.js",
+    "prod": "node node_modules/webpack/bin/webpack.js --config webpack.config.vendor.js --env.prod && node node_modules/webpack/bin/webpack.js --env.prod"
   },
   "devDependencies": {
-    "@angular/animations": "4.2.5",
-    "@angular/common": "4.2.5",
-    "@angular/compiler": "4.2.5",
-    "@angular/compiler-cli": "4.2.5",
-    "@angular/core": "4.2.5",
-    "@angular/forms": "4.2.5",
-    "@angular/http": "4.2.5",
-    "@angular/platform-browser": "4.2.5",
-    "@angular/platform-browser-dynamic": "4.2.5",
-    "@angular/platform-server": "4.2.5",
-    "@angular/router": "4.2.5",
-    "@ngtools/webpack": "1.5.0",
+    "@angular/animations": "6.1.10",
+    "@angular/cdk": "^6.4.7",
+    "@angular/cli": "^6.2.5",
+    "@angular/common": "6.1.10",
+    "@angular/compiler": "6.1.10",
+    "@angular/compiler-cli": "6.1.10",
+    "@angular/core": "6.1.10",
+    "@angular/forms": "6.1.10",
+    "@angular/http": "6.1.10",
+    "@angular/material": "^6.4.7",
+    "@angular/platform-browser": "6.1.10",
+    "@angular/platform-browser-dynamic": "6.1.10",
+    "@angular/platform-server": "6.1.10",
+    "@angular/router": "6.1.10",
+    "@ngtools/webpack": "^6.2.5",
     "@types/chai": "4.0.1",
-    "@types/jasmine": "2.5.53",
+    "@types/jasmine": "^2.5.53",
     "@types/webpack-env": "1.13.0",
     "angular2-router-loader": "0.3.5",
     "angular2-template-loader": "0.6.2",
     "aspnet-prerendering": "^3.0.1",
-    "aspnet-webpack": "^2.0.1",
-    "awesome-typescript-loader": "3.2.1",
-    "bootstrap": "3.3.7",
+    "aspnet-webpack": "^2.0.3",
+    "awesome-typescript-loader": "^3.2.1",
     "chai": "4.0.2",
+    "codelyzer": "^4.5.0",
     "css": "2.2.1",
     "css-loader": "0.28.4",
     "es6-shim": "0.35.3",
     "event-source-polyfill": "0.0.9",
     "expose-loader": "0.7.3",
-    "extract-text-webpack-plugin": "2.1.2",
+    "extract-text-webpack-plugin": "^2.1.2",
     "file-loader": "0.11.2",
     "html-loader": "0.4.5",
     "isomorphic-fetch": "2.2.1",
     "jasmine-core": "2.6.4",
-    "jquery": "3.2.1",
     "json-loader": "0.5.4",
     "karma": "1.7.0",
     "karma-chai": "0.1.0",
@@ -49,14 +54,17 @@
     "preboot": "4.5.2",
     "raw-loader": "0.5.1",
     "reflect-metadata": "0.1.10",
-    "rxjs": "5.4.2",
+    "rxjs": "6.3.3",
+    "rxjs-tslint": "^0.1.5",
     "style-loader": "0.18.2",
     "to-string-loader": "1.1.5",
-    "typescript": "2.4.1",
+    "tslint": "^5.11.0",
+    "typescript": "2.9.2",
     "url-loader": "0.5.9",
-    "webpack": "2.5.1",
+    "webpack": "^2.5.1",
     "webpack-hot-middleware": "2.18.2",
     "webpack-merge": "4.1.0",
-    "zone.js": "0.8.12"
-  }
+    "zone.js": "0.8.26"
+  },
+  "dependencies": {}
 }
