@@ -1,21 +1,24 @@
-import { NgModule } from '@angular/core';
-import { BrowserModule } from '@angular/platform-browser';
-import { AppModuleShared } from './app.shared.module';
-import { AppComponent } from './components/app/app.component';
+import { NgModule } from "@angular/core";
+import { BrowserModule } from "@angular/platform-browser";
+import { BrowserAnimationsModule } from "@angular/platform-browser/animations";
+
+import { AppComponent } from "./app.component";
+import { AppModuleShared } from "./app.shared.module";
 
 @NgModule({
-    bootstrap: [ AppComponent ],
+    bootstrap: [AppComponent],
     imports: [
         BrowserModule,
-        AppModuleShared
+        BrowserAnimationsModule,
+        AppModuleShared,
     ],
     providers: [
-        { provide: 'BASE_URL', useFactory: getBaseUrl }
-    ]
+        { provide: "BASE_URL", useFactory: getBaseUrl },
+    ],
 })
 export class AppModule {
 }
 
 export function getBaseUrl() {
-    return document.getElementsByTagName('base')[0].href;
+    return document.getElementsByTagName("base")[0].href;
 }
