@@ -1,9 +1,9 @@
-import { Component } from '@angular/core';
+import { Component } from "@angular/core";
 
 @Component({
-    selector: 'nav-menu',
-    templateUrl: './navmenu.component.html',
-    styleUrls: ['./navmenu.component.css']
+    selector: "mtp-menu",
+    styleUrls: ["./navmenu.component.css"],
+    templateUrl: "./navmenu.component.html",
 })
 export class NavMenuComponent {
 }
