@@ -3,95 +3,102 @@ const webpack = require("webpack");
 const merge = require("webpack-merge");
 const AotPlugin = require("@ngtools/webpack").AotPlugin;
 const CheckerPlugin = require("awesome-typescript-loader").CheckerPlugin;
+const TsConfigPathsPlugin = require('awesome-typescript-loader').TsConfigPathsPlugin;
 
 module.exports = (env) => {
-  // Configuration in common to both client-side and server-side bundles
-  const isDevBuild = !(env && env.prod);
-  const sharedConfig = {
-      stats: { modules: false },
-      context: __dirname,
-      resolve: { extensions: [".js", ".ts"] },
-      output: {
-          filename: "[name].js",
-          publicPath: "dist/" // Webpack dev middleware, if enabled, handles requests for this URL prefix
-        },
-      module: {
-          rules: [
-              {
-                test: /\.ts$/,
-                use: isDevBuild
-                       ? ["awesome-typescript-loader?silent=true", "angular2-template-loader", "angular2-router-loader"]
-                       : "@ngtools/webpack"
-              },
-              { test: /\.html$/, use: "html-loader?minimize=false" },
-              { test: /\.css$/, use: ["to-string-loader", isDevBuild ? "css-loader" : "css-loader?minimize"] },
-              { test: /\.(png|jpg|jpeg|gif|svg)$/, use: "url-loader?limit=25000" }
+    // Configuration in common to both client-side and server-side bundles
+    const isDevBuild = !(env && env.prod);
+    const sharedConfig = {
+        stats: {modules: false},
+        context: __dirname,
+        resolve: {
+            extensions: ['.js', '.ts'],
+            plugins: [
+                new TsConfigPathsPlugin()
             ]
         },
-      plugins: [new CheckerPlugin()]
+
+        output: {
+            filename: "[name].js",
+            publicPath: "dist/" // Webpack dev middleware, if enabled, handles requests for this URL prefix
+        },
+        module: {
+            rules: [
+                {
+                    test: /\.ts$/,
+                    use: isDevBuild
+                        ? ["awesome-typescript-loader?silent=true", "angular2-template-loader", "angular2-router-loader"]
+                        : "@ngtools/webpack"
+                },
+                {test: /\.html$/, use: "html-loader?minimize=false"},
+                {test: /\.css$/, use: ["to-string-loader", isDevBuild ? "css-loader" : "css-loader?minimize"]},
+                {test: /\.(png|jpg|jpeg|gif|svg)$/, use: "url-loader?limit=25000"}
+            ]
+        },
+        plugins: [new CheckerPlugin()]
     };
 
-  // Configuration for client-side bundle suitable for running in browsers
-  const clientBundleOutputDir = "./wwwroot/dist";
-  const clientBundleConfig = merge(sharedConfig, {
-      entry: { 'main-client': "./ClientApp/boot.browser.ts" },
-      output: { path: path.join(__dirname, clientBundleOutputDir) },
-      plugins: [
-          new webpack.DllReferencePlugin({
-              context: __dirname,
-              manifest: require("./wwwroot/dist/vendor-manifest.json")
+    // Configuration for client-side bundle suitable for running in browsers
+    const clientBundleOutputDir = "./wwwroot/dist";
+    const clientBundleConfig = merge(sharedConfig, {
+        entry: {'main-client': "./ClientApp/boot.browser.ts"},
+        output: {path: path.join(__dirname, clientBundleOutputDir)},
+        plugins: [
+            new webpack.DllReferencePlugin({
+                context: __dirname,
+                manifest: require("./wwwroot/dist/vendor-manifest.json")
             })
         ].concat(isDevBuild
-                 ? [
-                   // Plugins that apply in development builds only
-                   new webpack.SourceMapDevToolPlugin({
-                       filename: "[file].map", // Remove this line if you prefer inline source maps
-                       moduleFilenameTemplate:
-                         path.relative(clientBundleOutputDir,
-                           "[resourcePath]") // Point sourcemap entries to the original file locations on disk
-                     })
-                 ]
-                 : [
-                   // Plugins that apply in production builds only
-                   new webpack.optimize.UglifyJsPlugin(),
-                   new AotPlugin({
-                       tsConfigPath: "./tsconfig.json",
-                       entryModule: path.join(__dirname, "ClientApp/app/app.browser.module#AppModule"),
-                       exclude: ["./**/*.server.ts"]
-                     })
-                 ])
+            ? [
+                // Plugins that apply in development builds only
+                new webpack.SourceMapDevToolPlugin({
+                    filename: "[file].map", // Remove this line if you prefer inline source maps
+                    moduleFilenameTemplate:
+                        path.relative(clientBundleOutputDir,
+                            "[resourcePath]") // Point sourcemap entries to the original file locations on disk
+                })
+            ]
+            : [
+                // Plugins that apply in production builds only
+                new webpack.optimize.UglifyJsPlugin(),
+                new AotPlugin({
+                    tsConfigPath: "./tsconfig.json",
+                    entryModule: path.join(__dirname, "ClientApp/app/app.browser.module#AppModule"),
+                    exclude: ["./**/*.server.ts"]
+                })
+            ])
     });
 
-  // Configuration for Electron bundle suitable for running in ElectronNet
-  const electronBundleOutputDir = "./obj/Host/bin/wwwroot/dist";
-  const electronBundleConfig = merge(sharedConfig, {
-      entry: { 'main-client': "./ClientApp/boot.browser.ts" },
-      output: { path: path.join(__dirname, electronBundleOutputDir) },
-      plugins: [
-          new webpack.DllReferencePlugin({
-              context: __dirname,
-              manifest: require("./wwwroot/dist/vendor-manifest.json")
+    // Configuration for Electron bundle suitable for running in ElectronNet
+    const electronBundleOutputDir = "./obj/Host/bin/wwwroot/dist";
+    const electronBundleConfig = merge(sharedConfig, {
+        entry: {'main-client': "./ClientApp/boot.browser.ts"},
+        output: {path: path.join(__dirname, electronBundleOutputDir)},
+        plugins: [
+            new webpack.DllReferencePlugin({
+                context: __dirname,
+                manifest: require("./wwwroot/dist/vendor-manifest.json")
             })
         ].concat(isDevBuild
-                 ? [
-                   // Plugins that apply in development builds only
-                   new webpack.SourceMapDevToolPlugin({
-                       filename: "[file].map", // Remove this line if you prefer inline source maps
-                       moduleFilenameTemplate:
-                         path.relative(electronBundleOutputDir,
-                           "[resourcePath]") // Point sourcemap entries to the original file locations on disk
-                     })
-                 ]
-                 : [
-                   // Plugins that apply in production builds only
-                   new webpack.optimize.UglifyJsPlugin(),
-                   new AotPlugin({
-                       tsConfigPath: "./tsconfig.json",
-                       entryModule: path.join(__dirname, "ClientApp/app/app.browser.module#AppModule"),
-                       exclude: ["./**/*.server.ts"]
-                     })
-                 ])
+            ? [
+                // Plugins that apply in development builds only
+                new webpack.SourceMapDevToolPlugin({
+                    filename: "[file].map", // Remove this line if you prefer inline source maps
+                    moduleFilenameTemplate:
+                        path.relative(electronBundleOutputDir,
+                            "[resourcePath]") // Point sourcemap entries to the original file locations on disk
+                })
+            ]
+            : [
+                // Plugins that apply in production builds only
+                new webpack.optimize.UglifyJsPlugin(),
+                new AotPlugin({
+                    tsConfigPath: "./tsconfig.json",
+                    entryModule: path.join(__dirname, "ClientApp/app/app.browser.module#AppModule"),
+                    exclude: ["./**/*.server.ts"]
+                })
+            ])
     });
 
-  return [clientBundleConfig, electronBundleConfig];
-};
+    return [clientBundleConfig, electronBundleConfig];
+}
