@@ -8,19 +8,18 @@ const treeShakableModules = [
     '@angular/compiler',
     '@angular/core',
     '@angular/forms',
-    '@angular/http',
+    '@angular/common/http',
     '@angular/platform-browser',
     '@angular/platform-browser-dynamic',
     '@angular/router',
     'zone.js',
+    '@angular/material',
 ];
 const nonTreeShakableModules = [
-    'bootstrap',
-    'bootstrap/dist/css/bootstrap.css',
     'es6-promise',
     'es6-shim',
     'event-source-polyfill',
-    'jquery',
+    '@angular/material/prebuilt-themes/indigo-pink.css',
 ];
 const allModules = treeShakableModules.concat(nonTreeShakableModules);
 
@@ -28,11 +27,11 @@ module.exports = (env) => {
     const extractCSS = new ExtractTextPlugin('vendor.css');
     const isDevBuild = !(env && env.prod);
     const sharedConfig = {
-        stats: { modules: false },
-        resolve: { extensions: [ '.js' ] },
+        stats: {modules: false},
+        resolve: {extensions: ['.js']},
         module: {
             rules: [
-                { test: /\.(png|woff|woff2|eot|ttf|svg)(\?|$)/, use: 'url-loader?limit=100000' }
+                {test: /\.(png|woff|woff2|eot|ttf|svg)(\?|$)/, use: 'url-loader?limit=100000'}
             ]
         },
         output: {
@@ -41,7 +40,6 @@ module.exports = (env) => {
             library: '[name]_[hash]'
         },
         plugins: [
-            new webpack.ProvidePlugin({ $: 'jquery', jQuery: 'jquery' }), // Maps these identifiers to the jQuery package (because Bootstrap expects it to be a global variable)
             new webpack.ContextReplacementPlugin(/\@angular\b.*\b(bundles|linker)/, path.join(__dirname, './ClientApp')), // Workaround for https://github.com/angular/angular/issues/11580
             new webpack.ContextReplacementPlugin(/angular(\\|\/)core(\\|\/)@angular/, path.join(__dirname, './ClientApp')), // Workaround for https://github.com/angular/angular/issues/14898
             new webpack.IgnorePlugin(/^vertx$/) // Workaround for https://github.com/stefanpenner/es6-promise/issues/100
@@ -54,10 +52,10 @@ module.exports = (env) => {
             // But for production builds, leave the tree-shakable ones out so the AOT compiler can produce a smaller bundle.
             vendor: isDevBuild ? allModules : nonTreeShakableModules
         },
-        output: { path: path.join(__dirname, 'wwwroot', 'dist') },
+        output: {path: path.join(__dirname, 'wwwroot', 'dist')},
         module: {
             rules: [
-                { test: /\.css(\?|$)/, use: extractCSS.extract({ use: isDevBuild ? 'css-loader' : 'css-loader?minimize' }) }
+                {test: /\.css(\?|$)/, use: extractCSS.extract({use: isDevBuild ? 'css-loader' : 'css-loader?minimize'})}
             ]
         },
         plugins: [
